[pattern-gen] Support const fields in encoding with total length of >32bit

+++ llvm/tools/pattern-gen/lib/InstrInfo.hpp
@@ -15,7 +15,7 @@ struct CDSLInstr
         uint8_t len;
         uint8_t dstOffset;
         uint8_t srcOffset;
-        uint32_t val;
+        uint64_t val;
     };
 
     enum FieldType
@@ -34,7 +34,7 @@ struct CDSLInstr
     struct Field
     {
         uint8_t len;
-        uint32_t constV;
+        uint64_t constV;
         std::string_view ident;
         uint32_t identIdx;
         FieldType type;
